fastapi_travel_AI/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
mysql-connector-python